Cache the constructed tree of entities
To speed up future retrieval, the built tree of entities is cached for
30 days. The built tree includes the parent-child-relationship, so no
further pre-processing is required.

The cached data is tagged with all of the entities' cache-tags, i.e., if
an entity or the bundle itself changes, the cache is invalidated, the
tree rebuilt, and the result cached again.