Workflow: add CI-definition
Run a style- and typecheck on each push, and on each PR.
For phpstan, this requires a full Drupal environment, configured with
drupal/entity_reference_tree and drupal/wisski as dependencies.

Figures out the module version of this thing by looking at the event:
for pull requests, it will check out a specific version; for normal
pushes, it will default to the main branch.