feat: make caching language aware
The cache key now takes the current user's language code into
consideration. It did not do this before, which resulted in
trees being cached non-deterministically based on whatever language was
set at the time of generation. If the language was switched after the
fact, the tree was not retrieved from cache in the new language, but the
one at generation time.